chore: moves LockedSelector
