+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2020-06-07T16:22:33.850Z\n"
-"PO-Revision-Date: 2020-06-07T16:22:33.850Z\n"
+"POT-Creation-Date: 2020-06-08T08:49:45.695Z\n"
+"PO-Revision-Date: 2020-06-08T08:49:45.695Z\n"
 
 msgid "Please provide a valid date"
 msgstr ""
@@ -510,12 +510,6 @@ msgid "New {{TET}}"
 msgstr ""
 
 msgid "Working Lists"
-msgstr ""
-
-msgid "Could not get organisation unit"
-msgstr ""
-
-msgid "Program doesn't exist"
 msgstr ""
 
 msgid ""
