+++ src/core_modules/capture-core/components/Pages/NewEvent/DataEntry/epics/newEventDataEntry.epics.js
@@ -37,7 +37,7 @@ import getStageForEventProgram from '../../../../../metaData/helpers/EventProgra
 import getDataEntryKey from '../../../../DataEntry/common/getDataEntryKey';
 import { getProgramFromProgramIdThrowIfNotFound, TrackerProgram } from '../../../../../metaData';
 import { actionTypes as crossPageActionTypes } from '../../../actions/crossPage.actions';
-import { lockedSelectorActionTypes } from '../../../components/LockedSelector/actions';
+import { lockedSelectorActionTypes } from '../../../../LockedSelector/actions';
 
 const errorMessages = {
     PROGRAM_OR_STAGE_NOT_FOUND: 'Program or stage not found',
