+++ src/core_modules/capture-core/components/Pages/ViewEvent/ViewEventPage.component.js
@@ -2,7 +2,7 @@
 import React from 'react';
 import ViewEvent from './ViewEventComponent/ViewEvent.container';
 import ViewEventNewRelationshipWrapper from './Relationship/ViewEventNewRelationshipWrapper.container';
-import { LockedSelector } from '../components/LockedSelector/container';
+import { LockedSelector } from '../../LockedSelector/container';
 
 type Props = {
   formInputInProgess: boolean,
